feat: pss-cli-work-adapter-001 - Own Work CLI Service construction over the Work root
Move durable Work list/show/move/visualize CLI behavior into a Work-owned adapter Service constructed from ListRequestPreparation and VisualizationOperation, leaving pkg/transports/cli/work as thin facades.
