+++ pkg/transports/cli/work/helpers_test.go
@@ -1,0 +1,8 @@
+package work
+
+func stringValue(value *string) string {
+	if value == nil {
+		return ""
+	}
+	return *value
+}
