+++ pkg/transports/cli/work/list_request_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"net/url"
 	"strings"
 	"testing"
 
@@ -33,55 +32,6 @@ func (prepare testListRequestPreparation) PrepareListRequest(
 		summary = prepare.FilterSummary
 	}
 	return workservice.PreparedListRequest{Options: options, FilterSummary: summary}, nil
-}
-
-func TestBuildListRequest_NormalizesQueryWithoutHTTP(t *testing.T) {
-	nextToken := encodeCursor("work/42")
-	request, err := buildListRequest(ListConfig{Context: context.Background(),
-		Server:       "https://factory.example",
-		SessionID:    "session/alpha",
-		StateName:    "in review",
-		StateType:    "PROCESSING",
-		Name:         "Plan & review",
-		WorkTypeName: "story/type",
-		TraceID:      "trace+1",
-		SortBy:       "state.type",
-		MaxResults:   25,
-		NextToken:    nextToken,
-	}, workservice.PreparedListRequest{
-		Options: workservice.ListOptions{
-			StateName: "in review", StateType: "PROCESSING", Name: "Plan & review",
-			WorkTypeName: "story/type", TraceID: "trace+1", SortBy: "state.type",
-			MaxResults: 25, NextToken: nextToken,
-		},
-		FilterSummary: "state.name,state.type,name,workTypeName,traceId,sortBy",
-	})
-	if err != nil {
-		t.Fatalf("buildListRequest: %v", err)
-	}
-
-	wantQuery := url.Values{
-		"state.name":   {"in review"},
-		"state.type":   {"PROCESSING"},
-		"name":         {"Plan & review"},
-		"workTypeName": {"story/type"},
-		"traceId":      {"trace+1"},
-		"sortBy":       {"state.type"},
-		"maxResults":   {"25"},
-		"nextToken":    {nextToken},
-	}
-	if got := request.endpoint.Path; got != "/factory-sessions/session/alpha/work" {
-		t.Fatalf("endpoint path = %q", got)
-	}
-	if got := request.endpoint.EscapedPath(); got != "/factory-sessions/session%2Falpha/work" {
-		t.Fatalf("escaped endpoint path = %q", got)
-	}
-	if got := request.endpoint.RawQuery; got != wantQuery.Encode() {
-		t.Fatalf("query = %q, want %q", got, wantQuery.Encode())
-	}
-	if got := request.filterSummary; got != "state.name,state.type,name,workTypeName,traceId,sortBy" {
-		t.Fatalf("filter summary = %q", got)
-	}
 }
 
 func TestBuildListRequest_RejectsInvalidQueryWithoutHTTP(t *testing.T) {
